refactor(site): move diff viewers back into AgentsPage
